refactor: move CodeEditor to shared components and remove circular dependency
