+++ src/editors/sharedComponents/CodeEditor/index.scss
