+++ src/editors/sharedComponents/CodeEditor/index.test.jsx
@@ -4,9 +4,10 @@ import { shallow } from 'enzyme';
 import { EditorState } from '@codemirror/state';
 import { EditorView } from '@codemirror/view';
 import { html } from '@codemirror/lang-html';
-import { formatMessage, MockUseState } from '../../../../../testUtils';
+import { formatMessage, MockUseState } from '../../../testUtils';
 import alphanumericMap from './constants';
 import * as module from './index';
+import hooks from './hooks';
 
 jest.mock('@codemirror/view');
 
@@ -32,7 +33,7 @@ jest.mock('codemirror', () => ({
   basicSetup: 'bAsiCSetUp',
 }));
 
-const state = new MockUseState(module.hooks);
+const state = new MockUseState(hooks);
 
 describe('CodeEditor', () => {
   describe('Hooks', () => {
@@ -45,7 +46,7 @@ describe('CodeEditor', () => {
         state.mock();
       });
       it('prepareShowBtnEscapeHTML', () => {
-        const hook = module.hooks.prepareShowBtnEscapeHTML();
+        const hook = hooks.prepareShowBtnEscapeHTML();
         expect(state.stateVals.showBtnEscapeHTML).toEqual(hook.showBtnEscapeHTML);
         hook.hideBtn();
         expect(state.setState.showBtnEscapeHTML).toHaveBeenCalledWith(false);
@@ -58,7 +59,7 @@ describe('CodeEditor', () => {
       const cleanText = `${Object.values(alphanumericMap).join(' , ')}`;
 
       it('escapes alphanumerics and sets them to be literals', () => {
-        expect(module.hooks.cleanHTML({ initialText: dirtyText })).toEqual(cleanText);
+        expect(hooks.cleanHTML({ initialText: dirtyText })).toEqual(cleanText);
       });
     });
 
@@ -79,7 +80,7 @@ describe('CodeEditor', () => {
       };
       const mockHideBtn = jest.fn();
       it('unescapes literals and sets them to be alphanumerics', () => {
-        module.hooks.escapeHTMLSpecialChars({ ref, hideBtn: mockHideBtn });
+        hooks.escapeHTMLSpecialChars({ ref, hideBtn: mockHideBtn });
         expect(mockDispatch).toHaveBeenCalled();
         expect(mockHideBtn).toHaveBeenCalled();
       });
@@ -96,7 +97,7 @@ describe('CodeEditor', () => {
         },
       };
       beforeEach(() => {
-        module.hooks.createCodeMirrorDomNode(props);
+        hooks.createCodeMirrorDomNode(props);
       });
       it('calls useEffect and sets up codemirror objects', () => {
         const [cb, prereqs] = React.useEffect.mock.calls[0];
@@ -120,16 +121,16 @@ describe('CodeEditor', () => {
           },
           value: 'mOcKhTmL',
         };
-        jest.spyOn(module.hooks, 'createCodeMirrorDomNode').mockImplementation(() => ({}));
+        jest.spyOn(hooks, 'createCodeMirrorDomNode').mockImplementation(() => ({}));
       });
       afterAll(() => {
         jest.clearAllMocks();
       });
       test('Renders and calls Hooks ', () => {
-        jest.spyOn(module.hooks, 'prepareShowBtnEscapeHTML').mockImplementation(() => ({ showBtnEscapeHTML: true, hideBtn: mockHideBtn }));
+        jest.spyOn(hooks, 'prepareShowBtnEscapeHTML').mockImplementation(() => ({ showBtnEscapeHTML: true, hideBtn: mockHideBtn }));
         // Note: ref won't show up as it is not acutaly a DOM attribute.
         expect(shallow(<module.CodeEditor {...props} />)).toMatchSnapshot();
-        expect(module.hooks.createCodeMirrorDomNode).toHaveBeenCalled();
+        expect(hooks.createCodeMirrorDomNode).toHaveBeenCalled();
       });
     });
   });
